Exclude mobile_distribution code from coverage analysis
