+++ back/setup.cfg
@@ -27,6 +27,7 @@ exclude_lines =
     except Exception:
 omit =
     */*main.py
+    */mobile_distribution/*
 
 [coverage:html]
 directory = back/htmlcov
